remove annotation_data_loader, rename scvi_data_loader to scvi_dataloader
